Introduce weather data cache
This patch introduces weather data cache. Before sending a request to
weather provider, the cache is checked, and if the corresponding weather
information exists in the cache and is not outdated, then it's used in
the application.

While working with GPS location, the cache does not perform a direct
comparison of coordinates. Instead it checks if the new coordinate is
in the specified area around the cached one, or not.

Task-number: QTBUG-60467
Change-Id: I150a1023587be656ee8b190d1c2bce667c6301e2
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit 0d908c344c80cfcadb5ea1bb964bd4d8332d8803)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>