Introduce WeatherApi backend
This patch introduces the weather backend from weatherapi.com and
provides the means to switch between backends at runtime. This will be
done automatically in case one of the backends does not respond.

As we currently use different backends, we had to update the QML part
that is responsible for showing the icons. We no longer use the icons
from web-site for the forecast, but use the same icon set as for the
current weather.
This required introducing methods to convert backend-specific weather
icon information to a common format used in the application.
While on it, a new type of weather icon was introduced (taken from the
same icon set).

Different weather backends provide weather forecast for different
amount of days (due to license restrictions), so the QML code now makes
use of the Repeater that dynamically changes the amount of shown days.

Task-number: QTBUG-60467
Change-Id: Ic9aa3a97ec440dddb38f06edfff8a8434724d118
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit 6458b9b60a016bbe0ad6574a1965c49dcd9383f3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>