CoordinateAnimation: fix direction interpolation
The interpolation for the West and East directions were actually
confused. When the CoordinateAnimaion.East direction was selected, the
interpolation was done in West direction and vice versa.
This patch fixes the issue.

[ChangeLog][CoordinateAnimation][Important Behavior Change] The value
of direction property for CoordinateAnimation is now handled correctly.
Previously the values were confused, so specifying
CoordinateAnimation.East direction was actually leading to moving West
and vice versa.

Pick-to: 6.2
Change-Id: I8e73fc4ad5cabd80bca38c624c70919e54c55127
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>