Doc: Correct a broken link about GPS's accuracy
The site containing the link has been down since at least 2017 and its
author seems to be enjoying his retirement.

As it is not expected that the site will be online again or that its
content will be moved to a different location, the link was
substituted with one from Internet Archive's The Wayback Machine.

To ease the reading of the paragraph, the `\l` command was given a
link-text argument referencing the content of the website page.

Task-number: QTBUG-96127
Pick-to: 6.2
Change-Id: I3ee26fa3bac795ae03a202c3de0e6c157b4ee074
Reviewed-by: Paul Wicking <paul.wicking@qt.io>