Remove QAbstractDeclarativeData::parentChanged
This method was removed from qtbase in
e8b7e4e96a23b87011abeb1e4f5f694654866e3f.
This commit removes the usage of this method from qtlocation
and attempts to achieve the same behavior by introducing a
custom setParentAndNotify() method which emits the
parentChanged() signal.
The code is also updated to make use of the new method.

This is part of a bigger work to port QtLocation maps to Qt6.

Task-number: QTBUG-96795
Change-Id: I4523cf567b4510fc17473a37c689c689ee53130e
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>