QGeoRoutingManager doc example into snippet file
QGeoRoutingManager documentation has and example code with full C++ class. This also contains Q_OBJECT macro. But this macro is checked by AutoMoc (doesn't care if it's in a comment) and gives build error that moc_ file is not included. So, moving this example code into snippet file to get over that.

This is part of a bigger work to port QtLocation maps to Qt6.

Task-number: QTBUG-96795
Change-Id: I581ed5df55a49ce761423eba3353b683f2e2409b
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>