QDeclarativeGeoMap: reorder new parameters application after camera capabilities change
When the camera capabilities are changed, we try to re-apply such
parameters as zoom, tilt and FOV. Each of this parameters has min and
max supported values.
Before this patch we first applied the maximum value, and then the
minimum. This was triggering an assert in the unit tests.
With this patch we will first apply the minimum value. It will help to
unblock the CI, but still may trigger asserts under some conditions.
So a more general fix is required.

Task-number: QTBUG-96795
Change-Id: I7671981377e5a3c4f47fdbd4cf05aa540581b45f
Reviewed-by: Lauri Laanmets <lauri.laanmets@eesti.ee>
Reviewed-by: Andreas Buhr <andreas.buhr@qt.io>