Use current touch position when switching from one to two point mode
When panning the map with one finger to an offset and then placing
second finger on the map, then map will not jump anymore because in
'startTwoTouchPoints()' we now use 'scenePosition()' instead of
'scenePressPosition()'.

When two touch points are introduced, 'start point' of both must be
reset and not be based on the very first touch point of the first
finger.

Task-number: QTBUG-96795
Change-Id: I846691ada9ecfb6d10c855be50be807cea2f24a6
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>