Clean up mix of polymorph and value semantics
QGeoManeuver had a protected constructor only to allow the test to
replace the private implementation, which is polymorph. There is no use
of this possibility to override the implementation of the class outside
the test.

If we want to bring this back, then a factory that the service provider
plugins implement and through which the framework creates the
implementations would be the typical architecture, without forcing the
subclassing of value types with the risk for slicing or memory leaks
when destructors are not virtual.

Pick-to: 6.2
Task-number: QTBUG-105206
Change-Id: I3358da3939c43dfb83c4b2b10a792044590263f7
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>