Modernize and clean up QGeoRoute(Leg)
Implement move semantics for QGeoRoute, make operators hidden friends,
and add noexcept to non-allocating functions.

Remove default-implemented constructors, operators, and destructor from
QGeoRouteLeg. That class would ideally go away so that we don't have to
subclass the QGeoRoute value type at all. The leg-data lives either way
in the QGeoRoutePrivate(Default), so making the design so that a route
can be a list of routes would remove the need for the Leg class.

However, that has repercussions to the QML types, tests, documentation
etc, and needs to be discussed further.

Pick-to: 6.2
Task-number: QTBUG-105206
Change-Id: I4c451a74fbf42e06b1580df46bbccca4b0de1723
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>