Gadgetize Map Value types
This might allow us to remove the respective declarative wrapper
classes. However, QDeclarativeGeoRoute also provides access to a list of
segments, the path (as a QJSValue), and the routeQuery, which are all
not stored as part of QGeoRoute.

Change-Id: I65a25597f7d5540fb3edf837b48de160b5dde419
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit 38e62d40dad767cef98de704de9561e5082e413d)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>