Remove QDeclarativeGeoManeuver
Register the QGeoManeuver type directly as the RouteManeuver
QML type, and use it in QDeclarativeGeoRouteSegment.

The type is now uncreatable. It wasn't in Qt 5.15, but since there is no
setter-API that is exposed to QML (QDeclarativeGeoRouteSegment::maneuver
is a read-only property tagged as CONSTANT), there is no point in
creating a RouteManeuver object in QML anyway. The QML tests can get a
default-created QGeoManeuever from an empty route segment.

Document the C++ properties via \property, move the QML type
documentation into the qgeomaneuver.cpp file, and unify the language.

Note: The QML type should be routeManeuver, as value types should start
with a lower-case character. If we make that change, it will be in a
separate, compatibility-breaking commit for all the value types.

Change-Id: I277e7011f2082e2da9e0141db9bad0a60a524c88
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit da10ee9b67aaf0c9218c2b891efa06b323866470)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>