Remove QDeclarativeGeoMapType wrapper
Register the QGeoMapType gadget directly as geoMapType (via a Foreign),
and adapt QDeclarativeGeoMap to use it, simplifying the code.

The type was uncreatable before, and is implicitly uncreatable as a
value type now as well.

QGeoMapType also needs to be registered as a namespace so that the
MapStyle enum is available in QML.

Pick-to: 6.2
Change-Id: Ibf0b26eb799005e4fabcbd5496d019da812c8d70
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>