Remove QDeclarativeGeoRouteSegment wrapper
Register the QGeoRouteSegment gadget type directly as the routeSegment
QML type (via a Foreign), and use it in QDeclarativeGeoRoute.

The type is now uncreatable. It wasn't in 5.15, but since there is no
setter-API that is exposed to QML, there is no reason to have it be a
creatable type. Adjust the QML test to declare a property of type
routeSegment instead.

Move the QML type documentation into QGeoRouteSegment, and document the
corresponding C++ properties via \property. Unify the language a bit,
and follow up on the rename from RouteSegment to routeSegment.

Pick-to: 6.2
Change-Id: I1ebae0c1a9d056aa59510a3ab539b1d8bce8c6c7
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>