Change QJSValue properties to QList<QGeoCoordinate/Rectangle>
The QML engine is able to operate on lists of gadgets, there is no need
for using private APIs to operate on QJSValue.

For the time being, this breaks a QML construct like

    path[0].longitude = 0

This no longer changes the value of path[0].latitude in place. Instead,
use

    var path0 = path[0]
    path0.longitude = ...
    path[0] = path0

This is consistent with other properties that have type list<gadget>, as
QML operates on copies of values, not on references. Adapt the test case
accordingly.

Since support for value-initializing properties of type list<gadget>
requires plumbing in the QML engine, and registration of conversion
routines from QVariantMap to QGeoCoordinate, augment the test.

Remove the now unnecessary toList/fromList conversion functions, and
the dependency to private QtQml libraries.

Fixes: QTBUG-105241
Change-Id: I8f248c457a6de27a3b2680bdc948c5683ebc7fa0
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>