Register QPlaceRatings as a value type
Make QPlaceUser a gadget, and register it as a value type with the QML
engine. Remove declarative wrapper, and consolidate the documentation.

To support QML tests setting ratings properties from literal values,
register a converter from QJSValue. That is temporary and will become
obsolete once the QML engine can property-initialize, and perhaps even
instantiate, value types.

Adjust the tests and remove tests that verify that the type behaves like
an object. This is not entirely complete (some tests still compare
Place.ratings with `null`, but those tests pass for now, and will be
changed in later commits when more types get converted in this way.

Task-number: QTBUG-106482
Change-Id: I3694df414bb87adcfda3dc4b88e5206d4272bb82
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>