Register QPlaceAttribute as a value type
Make QPlaceAttribute a gadget, and register it as a value type with the
QML engine. Remove declarative wrapper, and consolidate the
documentation. Keep documentation for the ExtendedAttributes QML type
together with the placeAttribute QML value type for now.

Adjust the tests and remove tests that verify that the type behaves like
an object. Initialize placeAttribute properties as a grouped property.

Task-number: QTBUG-106482
Change-Id: Ia514ba4b2ae7f7af56dc0bcee19dca6dddbfd9cf
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>