Register QPlaceIcon as a QML value type
Make QPlaceIcon a gadget, and register it as a value type with the QML
engine. Remove declarative wrapper, and consolidate the documentation.

Adjust the tests and remove tests that verify that the type behaves like
an object. Initialize placeIcon properties as a grouped property, or via
a converter from QJSValue. That converter is a temporary solution to
enable tests, and can hopefully be removed once the QML engine supports
property-initialization of value types.

Task-number: QTBUG-106482
Change-Id: I2aa29102ee232afb3d71396bcd3288e2f99c0fc6
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>