Register QPlaceSupplier as a QML value type
Make QPlaceSupplier a gadget, and register it as a value type with the
QML engine. Remove declarative wrapper, and consolidate the
documentation.

Adjust the tests and remove tests that verify that the type behaves like
an object. Initialize placeSupplier properties as a grouped property, or
via a converter from QJSValue.

Change-Id: Id35ac1d9f03ce831013e4a7231302abd7b6cd7c8
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>