Register QPlaceContactDetail as a QML value type
Make QPlaceContactDetail a gadget, and register it as a value type with
the QML engine. Remove declarative wrapper, and consolidate the
documentation.

Adjust the tests and remove tests that verify that the type behaves like
an object. Initialize contactDetail properties as a grouped property.

The QDeclarativeContactDetails implementation of QQmlPropertyMap has to
stay, so rename the sources accordingly.

Change-Id: I473d29405963bbe594cd5f26168a237c189d1d3e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>