Cleanup: de-virtualize QGeoRouteSegmentPrivate
The possibility to override this type in plugins is not used anywhere,
and overcomplicates the code. There isn't even a way to create a
QGeoRouteSegment with a reimplementation of the private.

So remove this; if this is really needed for anything, then we can bring
it back later.

Change-Id: I34ad55d0ddf29cafc53e90e2a987c3faf904b437
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit 69e0dea5de987d5b6e684b721b322b0495921981)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>