Cleanup: de-virtualize QGeoManeuverPrivate
The possibility to override this type in plugins is not used anywhere,
and overcomplicates the code. There isn't even a way to create a
QGeoManeuver with a reimplementation of the private.

So remove this; if this is really needed for anything, then we can bring
it back later.

Change-Id: I4edf78a21f53593e609f6e86e04464d4e974b0df
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 87eb1a45b83e6044d7d0c83f98f419c8fbd5e007)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>