Cleanup: de-virtualize QPlaceAttributePrivate
The possibility to override this type in plugins is not used anywhere,
and overcomplicates the code. There isn't even a way to create a
QPlaceAttribute with a reimplementation of the private.

So remove this; if this is really needed for anything, then we can bring
it back later.

Change-Id: I4881583b87b775a4bcaf36bc972248b321e0ef2b
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
(cherry picked from commit 29bf63108f954119a0df6c70dd283367777a7e23)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>