Cleanup: de-virtualize QGeoRoutePrivate
The possibility to override this type in plugins was used for metadata
that might just as well live in the extended attributes. And the usecase
of making the basic passing of data from a private through a public type
does not justify the overhead, and can be solved by simply initializing
the value in the test plugin through the public setter.

So remove this; if this is really needed for anything, then we can bring
it back later.

Change-Id: Iad8689b54f64fa89535d0ebc3c399b4fd054f9d7
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>