Cleanup: de-virtualize QGeoCodeReplyPrivate
The only call to the single virtual function, extraData(), was commented
out. The subclass in the OSM plugin only wrote extra data for debug runs
and the subclass in the mock plugin is never used to test anything anyway.

So remove all this overhead. As a drive-by, refactor to member
initialization and use qsizetype for members that refer to indices in a
QList.

Change-Id: I600b4637bc367ae45d51b2a7f9a3ae0b78fb77c4
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit ff4207b1a41db4d23d4d0183152776e73f252a3a)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>