Cleanup: Remove duplicate storage of values from declarative wrappers
QPlace and QPlaceCategory already store the icon, ratings, supplier etc
as values. The declarative wrappers have to emit signals when those
change, but that can only happen when either the QPlaceCategory or
QPlace itself changes, or when the individual sub-properties are changed
through the setters of the declarative type.

So we can remove the double-book-keeping and treat those value types
like all other values (e.g. QPlace::name).

Change-Id: I2703529ee625e23937b3d12ec57e60500f8e1e62
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>