Remove MapObjects from labs
The feature was designed to enable map backends to implement their own
rendering of the map, and map items. It was never really used by any of
the backends we supported, and introduced a lot of complexity to the
code base.

The idea as such has perhaps merit, but for now we focus on the basic
functionality of Qt Location. Support for Qt Location when the Qt
Quick scene graph is not used is not a part of that.

Change-Id: I08e460043a0bd2600bad79b6ce2bb18e40e19eb6
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>