Build system: move to declarative type registration
Adjust the build system accordingly, attaching a QML module with
the respective implementations to the QtLocation target.

An explicit function is still needed to register converters, and will
be removed once the improved support for value types has merged.

This should enable the qml compiler to generate C++ code, which then
requires linking against the library.

The MapParameter type was renamed to DynamicMapParamter as per the
5.11 change log, so remove the MapParameter alias now and adjust
tests.

Task-number: QTBUG-106886
Change-Id: Id8765c1bc3b98d447d768b246b21a16f71bfdf74
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>