Remove QDeclarativeGeoWaypoint
The type was introduced as a QML-wrapper around QGeoCoordinate, which is
now a registered value type for QML. QGeoCoordindate doesn't have a
bearing property, but as waypoints could be initialized with a list of
QGeoCoordinate anyway, 'bearing' was only inconsistently supported. If
we want to bring that back, then it might be best to add an optional
bearing to QGeoCoordinate in Qt Positioning.

Change-Id: I3d8cb64f3552dc5580945b17c3a9092a7205b9b8
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>