PositionSource: fix lastKnownPosition request at startup
As mentioned in QTBUG-107580, the m_active property is always false by
the time we invoke the tryAttach() method. That is because calling
setActive(true), even at construct time, does not update the value
immediately, but instead defers the update until the real position
source is created.

This patch fixes it by introducing an m_startPending flag and
checking it instead of m_active. This flag is set when
setActive(true) is called, so it will have the correct value in
tryAttach().

This commit amends 53388508305ada6c364919042d5bd90ab5611337.

Fixes: QTBUG-107580
Task-number: QTBUG-103478
Change-Id: I72d532205f4bc1387dad0bd0e65a1871a840b356
Reviewed-by: Juha Vuolle <juha.vuolle@insta.fi>