Simplify by removing the LineStrip mode for MapPolyline
Have two modes (Software, OpenGL) for polyline, as it is the case
for all other items (rectangle, circle, polygon). OpenGLLineStrip
is removed completely, whereas the enum value OpenGL is added with
the same value as OpenGLExtruded for symmetry with other items.

Drawing lines and expecting wide line (width > 1) support to be
avilable is highly non-portable: Direct 3D, Metal, and core profile
OpenGL contexts will not support widths other than 1, whereas with
Vulkan wide lines are an optional feature so it may or may not work.

As the 'backend' property is already marked as internal, it won't
present any consequences when it comes to the public API.

Internally this requires some untangling of the somewhat
intertwined node and material implementations.

Change-Id: I175ddb5f84128ed4d0fcf2939272e631566ff327
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>