Remove "OpenGL" map item implementations
The aim is to have a single solution. And due to certain cases
falling back to the default ("CPU") path, this "OpenGL" path is
not complete on its own. In addition, some rendering features
are not implemented (caps and joins), meaning additional work
is needed still to be on par with the default item rendering
path.

Change-Id: Ie67c8b5ac002c6660f3808c10d9fd9276a16e47e
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>