Update location examples to work with MapView/Handlers
MapView replaces Map in the examples to reintroduce event handling.
Some of the functionality is still in mapview.map and the code has to be
adjusted respectivly.

The following examples are affected: mapviewer, places, places_map.

The mapviewer example has a new menu to draw mapitems.

The places example draws places icons instead of a fixed marker.

Fixes: QTBUG-111005
Change-Id: I15ca2b3d3ecb7c35684b9fd4c2b7a86249add11c
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>