Correct and improve the rendering of QuickMapItems
Various MapItems were not rendered correctly, especially in corner cases.
This change ensures that MapItems are rendered correctly in the vast
majority of cases.

All MapItems are shown correctly if they wrap around the globe and
appear twice on the map.

Circles that span around the globe or are located near poles are shown
correclty and filled all the way to the border of the map.

Polygons are shown correctly including their holes.

The code was simplified and some artefacts of previous implementations
were removed.

Fixes: QTBUG-110701
Fixes: QTBUG-110511
Pick-to: 6.5
Change-Id: I1110659989436cd5a93f6ec26f75caa06d5f2b71
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>