Improve compatibility between OSM plugin and places example
The categories are searched as well as possible. However, searches for
e.g. all shops are not possible with OSM since a recent API change.

The category is written into the PlacesReply object so that they can be
displayed in the ui.

Change-Id: Ic08e5bef393313672775b6f39d365484ee8b0c58
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 992f57fcf993b9804dcda9b12bf4d67ed102f284)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>