Polygons/lines can now be rendered following the shortest path on the globe
This is enabled by interpolating the lines of polygons and paths.
The interpolating is done following the greater circle navigation and the
connection between corners of the polygon appear curved on the projected map.

This behavior can be turned on by setting a new property, called
referenceSurface. It can be set to ReferenceSurface.Map, drawing paths as lines on
the map or to ReferenceSurface.Globe, drawing path on the globe leading to
curves on the map. It is set to ReferenceSurface.Map on default, reproducing
the old implementation for polygons, polylines and rectangles.

The circle item was already using the great circle path before this
change. Its standard implementation was changed to draw a circle in map
coordinates with approximated radius. This should be sufficient for many
cases. To get the old implementation, referenceSurface has to be set to
ReferenceSurface.Map.

Fixes: QTBUG-94785
Change-Id: Ifdd1597a7116c3d220462f063656b04becb6422f
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>